info saving in the db seems to work now.  FREEKING AWESOME.
Adventurous types should check it out for day to day usage, just
keep in mind that the db format isn't set in stone yet.  In fact,
to make this work I had to add a decoder_type field to post_file -
so you will need to delete any already existing newsgroup databases
in order to use this correctly

darcs-hash:20060125190408-b89eb-aa4556e9dcc01a51ef6e5d8d653a2e0d4feb442e.gz

+++ src/base/database.cpp
@@ -5,6 +5,9 @@
 #include <unistd.h>
 #include <sys/types.h>
 #include <sys/stat.h>
+#include <sstream>
+
+using std::stringstream;
 
 sqlite3 *get_main_db(void)
 {
@@ -95,25 +98,33 @@ void restore_postsets_from_db(sqlite3 *db, NewsGroup *group)
 void restore_postfiles_from_db(sqlite3 *db, PostSet *set)
 {
     sqlite3_stmt *s;
-    string stmt = "SELECT postfile_no, num_pieces, name FROM post_files WHERE postset_no = ?";
+    string stmt = "SELECT postfile_no, num_pieces, name, decoder_type FROM post_files WHERE postset_no = ?";
     sqlite3_prepare(db, stmt.c_str(), stmt.length(), &s, 0);
     sqlite3_bind_int(s, 1, set->db_index); 
     while (SQLITE_ROW == sqlite3_step(s)){
         PostFile *file = set->file((char *)sqlite3_column_text(s, 2));
         file->db_index = sqlite3_column_int(s, 0);
         file->_num_file_pieces = sqlite3_column_int(s, 1);
+        file->decoder_type = (DecoderType)sqlite3_column_int(s, 3);
     }
     sqlite3_finalize(s);
 }
         
 void restore_ids_from_db(sqlite3 *db, PostFile *file)
 {
+    if(!file->db_index)
+        return;
+
     sqlite3_stmt *s;
     string stmt = "SELECT msg_id, status FROM file_pieces WHERE postfile_no = ? ORDER BY file_piece_no";
     sqlite3_prepare(db, stmt.c_str(), stmt.length(), &s, 0);
     sqlite3_bind_int(s, 1, file->db_index); 
     while (SQLITE_ROW == sqlite3_step(s)){
-        file->pieces.push_back(new FilePiece(sqlite3_column_int(s, 1), (PIECE_STATUS) sqlite3_column_int(s, 2), file));
+        FilePiece *new_piece = new FilePiece(sqlite3_column_int(s, 0), (PIECE_STATUS) sqlite3_column_int(s, 1), file);
+        file->pieces.push_back(new_piece);
+        stringstream st;
+        st << "Restoring piece " << new_piece->msg_id << " with status " << new_piece->status;
+        console->log(st.str());
     }
     file->has_db_pieces = 1;
     sqlite3_finalize(s);
@@ -151,7 +162,7 @@ void setup_postset_tables(sqlite3 *db)
 {
     vector<string> queries;
     queries.push_back("CREATE TABLE post_sets   (postset_no    INTEGER PRIMARY KEY, name VARCHAR)");
-    queries.push_back("CREATE TABLE post_files  (postfile_no   INTEGER PRIMARY KEY, postset_no INTEGER, num_pieces INTEGER, name VARCHAR)");
+    queries.push_back("CREATE TABLE post_files  (postfile_no   INTEGER PRIMARY KEY, postset_no INTEGER, num_pieces INTEGER, name VARCHAR, decoder_type INTEGER)");
     queries.push_back("CREATE TABLE file_pieces (file_piece_no INTEGER PRIMARY KEY, postfile_no INTEGER, status INTEGER, msg_id INTEGER)");
     run_queries(db, queries);
 }
@@ -194,7 +205,7 @@ void save_subscribed_groups_to_db(sqlite3* db)
 void save_postfiles(sqlite3 *db, PostSet *set)
 {
     sqlite3_stmt *pf;
-    string pf_stmt = "INSERT INTO post_files VALUES(NULL, ?, ?, ?)";
+    string pf_stmt = "INSERT INTO post_files VALUES(NULL, ?, ?, ?, ?)";
     sqlite3_prepare(db, pf_stmt.c_str(), pf_stmt.length(), &pf, 0);
     Uint32 max_no = set->files.size();
     for(Uint32 i=0; i<max_no; i++){
@@ -203,17 +214,22 @@ void save_postfiles(sqlite3 *db, PostSet *set)
             sqlite3_bind_int(pf, 1, set->db_index); 
             sqlite3_bind_int(pf, 2, file->num_pieces()); 
             sqlite3_bind_text(pf, 3, file->filename.c_str(), file->filename.length(), NULL);
+            sqlite3_bind_int(pf, 4, file->decoder_type);
             sqlite3_step(pf);
             sqlite3_reset(pf);
             file->db_index = sqlite3_last_insert_rowid(db);
         }
-        save_ids_to_db(db, file);
+        if(file->has_db_pieces)
+            save_ids_to_db(db, file);
     }
     sqlite3_finalize(pf);
 }
 
 void save_ids_to_db(sqlite3* db, PostFile *file)
 {
+    if(!file->has_db_pieces)
+        return;
+
     sqlite3_stmt *fp;
     string del_stmt = "DELETE from file_pieces where postfile_no = ?";
     sqlite3_prepare(db, del_stmt.c_str(), del_stmt.length(), &fp, 0);
@@ -237,9 +253,12 @@ void save_ids_to_db(sqlite3* db, PostFile *file)
 
 void restore_ids_from_db(PostFile *file)
 {
-    if(!file->db_index || file->has_db_pieces)
-    {
+    if(!file->db_index){
         file->has_db_pieces = 1;
+        return;
+    }
+    if(file->has_db_pieces)
+    {
         return;
     }
     sqlite3 *db = db_for_newsgroup(file->post_set->group);
